Rename convertor.py to metadata_convertor.py
