+++ dtp/__init__.py
@@ -1,6 +1,6 @@
 from dtp.core.downloader import Downloader
 from dtp.gen3.auth import Auth
-from dtp.gen3.convertor import MetadataConvertor
+from dtp.gen3.metadata_convertor import MetadataConvertor
 from dtp.gen3.exporter import Exporter
 from dtp.gen3.querier import Querier
 from dtp.gen3.submitter import Gen3Submitter
